[FEAT] 주문 컨트롤러 작성

+++ A server/src/main/java/backend/resource/order/dto/OrderRequestDto.java
@@ -2,12 +2,14 @@ package backend.resource.order.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class OrderRequestDto {
     private long productId; // 상품
-    private Double quantity; // 수량
+    private BigDecimal quantity; // 수량
     private String address; // 주소
 }
